saved restaurants component fully styled and functional
