+++ react-app/src/components/RestaurantProfile/RestaurantProfile.css
@@ -1,3 +1,8 @@
+.restaurant-profile-main-container {
+    height: auto;
+    overflow: scroll;
+}
+
 .restaurant-profile-image {
     width: 100%;
     height: 460px;
@@ -18,14 +23,9 @@
 
 .restaurant-profile-saved-restaurant-icon {
     height: 23px;
-    width: 17px;
+    width: 17.5px;
     margin-right: 13px;
 }
-
-/* .restaurant-profile-save-restaurant-icon-text-container {
-    display: flex;
-    position: relative;
-} */
 
 .restaurant-profile-save-restaurant-button {
     position: absolute;
@@ -117,6 +117,7 @@
 
 .restaurant-profile-lower {
     height: 1200px;
+    height: auto;
     margin-left: 30px;
 }
 
@@ -125,6 +126,7 @@
     display: flex;
     flex-direction: row;
     justify-content:center;
+    height: auto;
 }
 
 .restaurant-profile-description-container {
@@ -145,7 +147,8 @@
 .restaurant-profile-left {
     width: 620px;
     margin-right: 50px;
-    overflow: scroll;
+    /* overflow: scroll; */
+    height: auto;
 }
 
 .restaurant-reviews-count-text {
